reporto-backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
pytesseract
pypdf
python-dotenv
google-generativeai
gradio
